requirements.txt
fdk
oci-cli
